// src/app/types/user.ts
export type User = {
  avatar_url: string | undefined;
  name: string | undefined;
  twitter_username?: string | undefined;
  bio?: string | undefined;
  followers: number | undefined;
  public_repos: number | undefined;
  login: string | undefined;
  location: string | undefined;
  company: string | undefined;
};